Extract SessionUrl into pulse.url
RenderSession only owns the URL; the slot map and URL↔Signal
effects live in their own module. SessionUrl takes a send
callback so it does not import RenderSession.
